add trip with its photos and hotels

+++ src/app/utilties/images/images.component.ts
@@ -14,7 +14,7 @@ export class ImagesComponent implements OnInit {
 
 
 
-  @Input() id:number;
+  @Input() id: number;
   @Input() slides: any[] = [];
   @Input() files: any[];
   @Input() shown: boolean;
@@ -22,7 +22,7 @@ export class ImagesComponent implements OnInit {
   @Input() fileName: string;
   @Output() submit = new EventEmitter();
   @Output() deleteImage = new EventEmitter();
-
+  @Output('imagesChanged') imagesChanged = new EventEmitter();
   @ViewChild('imageModel') private imageModel;
   urls = new Array<string>();
   uploadPath: string;
@@ -37,6 +37,7 @@ export class ImagesComponent implements OnInit {
   }
 
   ngOnInit() {
+    console.log(this.slides);
     this.uploadPath = environment.upload + "images/" + this.fileName + "/";
   };
 
@@ -49,12 +50,31 @@ export class ImagesComponent implements OnInit {
       if (this.imagesService.checkImagesType(files[i].type))
         this.imagesToSend.append('image', files[i], files[i].name);
     }
-    this.save();
+    if (this.id) {
+      this.save();
+    } else {
+      this.getImagesNames();
+    }
+  }
+
+
+  getImagesNames() {
+    this.imagesService.getImagesNames(this.fileName, this.imagesToSend)
+      .subscribe(images => {
+        if (images['success']) {
+          this.tostr.success('it was added successfuly');
+          images['data'].forEach(image => {
+            this.slides.push(image);
+            this.imagesChanged.emit(this.slides);
+          })
+          this.imagesToSend = new FormData();
+        }
+      })
   }
 
 
   save() {
-    let add = this.imagesService.addImages(this.fileName, this.imagesToSend,this.id)
+    let add = this.imagesService.addImages(this.fileName, this.imagesToSend, this.id)
       .subscribe(images => {
         if (images['success']) {
           this.tostr.success('it was added successfuly', 'success');
@@ -70,12 +90,17 @@ export class ImagesComponent implements OnInit {
 
 
   removeSlide(index?: number): void {
+    if (!this.id) {
+      this.slides.splice(this.activeSlideIndex, 1);
+      this.imagesChanged.emit(this.slides);
+      return;
+    }
     let imageId = this.slides[this.activeSlideIndex]['id'];
     let delete1 = this.imagesService.deleteImage(this.fileName, imageId)
       .subscribe(data => {
         if (data['success']) {
           this.tostr.success('it was deleted successfuly', 'success');
-          this.slides.splice(this.activeSlideIndex,1);
+          this.slides.splice(this.activeSlideIndex, 1);
         }
       }, err => {
         this.tostr.error('there was a problem', 'error');
